% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelrunnR-package.R
\docType{package}
\name{modelrunnR-package}
\alias{modelrunnR}
\alias{modelrunnR-package}
\title{modelrunnR: Orchestrate R and Python Model Runs}
\description{
Orchestrates R and Python model runs via lightweight configuration and a common harness interface. Inspired by the orchestration layer of the panelmodeler package.
}
\author{
\strong{Maintainer}: Jesse Brandt \email{jesse.a.brandt@gmail.com}

Authors:
\itemize{
  \item Jesse Brandt \email{jesse.a.brandt@gmail.com}
}

}
\keyword{internal}
